requirements.txt
Django==1.9.6
django-filter==0.13.0
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
PyJWT==1.4.0
PyMySQL==0.7.4
